build specifically with FreeRTOS 10.6.2
Before this small change, the PicoHTTPServer project would try to build
against the development version of FreeRTOS (the tip of the master
branch).  At the current time this branch is in flux and PicoHTTPServer
does not build.

This commit specifies the exact version of FreeRTOS to build against as
10.6.2 (the most recent stable version), which I've verified works.

A better solution might be to use FreeRTOS as a submodule, so that
the PicoHTTPServer repo can explicitly control the FreeRTOS version,
instead of relying on the build script to select one.